CSES DP dice combinations prob

+++ DP/dice_combo.cpp
@@ -3,32 +3,32 @@ using namespace std;
 
 static const int MOD = 1e9 + 7;
 
-int f(int sum , vector<int> &dp) {
+long long f(int sum , vector<long long> &dp) {
     if(sum == 0) return 1;
     if(dp[sum] != -1) return dp[sum];
     // if i get a number 1
-    int ans = 0;
-    int getOne = 0;
+    long long ans = 0;
+    long long getOne = 0;
     if(sum - 1 >= 0) {
         getOne = f(sum - 1 , dp) % MOD;
     }
-    int getTwo = 0;
+    long long getTwo = 0;
     if(sum - 2 >= 0) {
         getTwo = f(sum - 2 , dp) % MOD;
     }
-    int getThree = 0;
+    long long getThree = 0;
     if(sum - 3 >= 0) {
         getThree = f(sum - 3 , dp) % MOD;
     }
-    int getFour = 0;
+    long long getFour = 0;
     if(sum - 4 >= 0) {
         getFour = f(sum - 4 , dp) % MOD;
     }
-    int getFive = 0;
+    long long getFive = 0;
     if(sum - 5 >= 0) {
         getFive = f(sum - 5 , dp) % MOD;
     }
-    int getSix = 0;
+    long long getSix = 0;
     if(sum - 6 >= 0) {
         getSix = f(sum - 6 , dp) % MOD;
     }
@@ -39,7 +39,7 @@ int f(int sum , vector<int> &dp) {
 int main() {
     int n;
     cin >> n;
-    vector<int> dp(n + 1 , -1);
+    vector<long long> dp(n + 1 , -1);
     int res = f(n , dp);
     cout << res << endl;
 }
